requirements.txt
-i https://pypi.org/simple
azure-functions==1.17.0
blinker==1.6.3; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
contourpy==1.1.1; python_version >= '3.8'
cycler==0.12.1; python_version >= '3.8'
flask==3.0.0
fonttools==4.43.1; python_version >= '3.8'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
kiwisolver==1.4.5; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2; python_version >= '3.7'
pillow==10.1.0
pillow-heif==0.13.1
pydicom==2.4.3
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
werkzeug==3.0.0; python_version >= '3.8'
